changing version to a stable release or so

+++ ports_table.c
@@ -2,6 +2,7 @@
 #include <linux/kernel.h>
 #include <linux/module.h>
 #include <linux/rbtree.h>
+#include <linux/types.h>
 
 #include "table_port.h"
 #include "portsDB.h"
@@ -11,11 +12,11 @@
 
 extern struct rb_root db;
 
-int insertPort(int port)
+int insertPort(u16 port)
 {
 	int ret;
 	struct portInfo *p = NULL;
-	printk(KERN_INFO "inserting port %d", port);
+	printk(KERN_INFO "inserting port %hu", port);
 
 	p = kmalloc(sizeof(*p),GFP_KERNEL);
 	p->port = port;
@@ -25,9 +26,9 @@ int insertPort(int port)
 	return -ENOTIMPLEMENTED;
 }
 
-int deletePort(int port)
+int deletePort(u16 port)
 {
-	printk(KERN_INFO "deleting port %d",port);
+	printk(KERN_INFO "deleting port %hu",port);
 	my_erase(&db,port);
 	return 0;
 }
